Squashed commit of the following:
commit 83436f3fc023b599d5b6298b1e6cf3792743e829
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Fri Jul 13 17:34:01 2018 +0200

    Fixbug for mode dev enable, reload button, fixbug save order product demat

commit 5d60a91ef605de8f8bf5a7cbcccd50f9b8fd045a
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Fri Jul 13 17:08:08 2018 +0200

    Rebase branch 1.x

commit 7517f2b7a0c3b123733f8add861aa158b435c1dd
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Fri Jul 13 16:43:40 2018 +0200

    Update message error for quantity product

commit e3b52f6709c745467cdf3e74f9bcbcebc48301a5
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Fri Jul 13 11:18:12 2018 +0200

    Bugfix: Rename post_code on postCodde

commit 1d406d60c0855709b6d16903275d15710ae13e1a
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Thu Jul 12 09:22:02 2018 +0200

    Change display name for module

commit 1e74e40f2f7a4d2f6db4932ebe88c62f5e10b4a5
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Wed Jul 11 17:44:13 2018 +0200

    Fix remove last item

commit b002beeb28f40eb240901e5a47033451a4104685
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Wed Jul 11 16:10:09 2018 +0200

    Fix sync cart & check customer logged for cart.estimate

commit 887156692ec762d0391081432611ecccdc2824d0
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Tue Jul 10 17:32:53 2018 +0200

    Fix carrier and display product get all products

commit 4ae157758f830db4b6fe849c69f6c05d79777caf
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Tue Jul 10 11:43:09 2018 +0200

    Fix url modal for 1.6 and add config url modal BO

commit 1c011b0f30643d424b585fd33d6b5811e93e490a
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Fri Jul 6 15:21:22 2018 +0200

    Check oneClickCustomization exist

commit 9693de0db5f6e39759ec5b52e5bb2526fbc4ab4c
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Fri Jul 6 11:37:37 2018 +0200

    Remove key url return

commit f8d18bd36257823078c90638e557db4f8ea1b845
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Fri Jul 6 10:21:38 2018 +0200

    Remove should ask stock for 1.5

commit f73c52a82890b22bf17e3e139f27dbb2037ade41
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Fri Jun 29 14:45:39 2018 +0200

    Remove manage stock

commit b18d825d7c95dd44b61baa4c192e2e70ae4cadf4
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Fri Jun 29 10:44:52 2018 +0200

    Change url cdn for environnement

commit 4049798ab81cd5c478b1fbc572bb5995a5bd70b4
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Wed Jun 27 11:48:17 2018 +0200

    Remove manage stock

commit 23837e993e0c8f288d37fab46fab330fe7af5090
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Tue Jun 26 12:01:19 2018 +0200

    Update set preload on click button oc

commit e9b3c165068708eb4f0f981d1fe82b84d60b3997
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Tue Jun 26 09:14:22 2018 +0200

    Remove id_cart on url confirmation

commit 9e25a8a532605eeae26c0467f919d16961fc6880
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Fri Jun 22 11:52:13 2018 +0200

    Change wording description

commit fc5603d810b8029e370774b90985a8fe03209c19
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Fri Jun 22 10:26:38 2018 +0200

    Remove secure_key for confirmation order OC

commit b8c8f8cfef52ba4e3962b446e2d0422d36bc4bd7
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Wed Jun 20 18:08:01 2018 +0200

    Change order load script

commit 419dbb50f20e57b94ed46ed4ab63ff9ec65ebb44
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Wed Jun 20 11:34:08 2018 +0200

    Update file upgrade && add remove conf for btn addr

commit 58003eddbe6c40adc641dfb808a1890da4aff52d
Author: Romain FEY <le_quartier@hotmail.fr>
Date:   Wed Jun 20 09:56:01 2018 +0200

    Add sync cart since product

commit 859dc34988b6bf465c2d4c9a18121b4a0aa05b73
Author: Romain FEY <romain.fey@pilipili-web.com>
Date:   Wed Apr 18 14:53:29 2018 +0200

    Update new lib

commit 46d4e9283751deea6870fc1eabf016aea738719c
Author: Romain FEY <romain.fey@pilipili-web.com>
Date:   Wed Apr 18 11:42:40 2018 +0200

    Add order_amount

+++ src/Service/InstallManager.php
@@ -244,6 +244,9 @@ class InstallManager
         PSConfiguration::deleteByName(Configuration::API_KEY_CUSTOM_ONECLICK);
         PSConfiguration::deleteByName(Configuration::ONE_CLICK_URL_CUSTOM);
         PSConfiguration::deleteByName(Configuration::API_ENV_ONECLICK);
+        PSConfiguration::deleteByName(Configuration::ONE_CLICK_MODAL_URL_PROD);
+        PSConfiguration::deleteByName(Configuration::ONE_CLICK_MODAL_URL_SANDBOX);
+        PSConfiguration::deleteByName(Configuration::ONE_CLICK_MODAL_URL_CUSTOM);
         // Conf custom 1-Click global
         PSConfiguration::deleteByName('FC_OYST_STATE_PAYMENT_ONECLICK');
         PSConfiguration::deleteByName('FC_OYST_BORDER_BTN');
@@ -312,6 +315,16 @@ class InstallManager
         PSConfiguration::deleteByName('FC_OYST_ID_BTN_PAYMENT');
         PSConfiguration::deleteByName('FC_OYST_ID_SMART_BTN_PAYMENT');
         PSConfiguration::deleteByName('FC_OYST_POSITION_BTN_PAYMENT');
+        // Conf 1-Click custom btn addr
+        PSConfiguration::deleteByName('FC_OYST_BTN_ADDR');
+        PSConfiguration::deleteByName('FC_OYST_WIDTH_BTN_ADDR');
+        PSConfiguration::deleteByName('FC_OYST_HEIGHT_BTN_ADDR');
+        PSConfiguration::deleteByName('FC_OYST_MARGIN_TOP_BTN_ADDR');
+        PSConfiguration::deleteByName('FC_OYST_MARGIN_LEFT_BTN_ADDR');
+        PSConfiguration::deleteByName('FC_OYST_MARGIN_RIGHT_BTN_ADDR');
+        PSConfiguration::deleteByName('FC_OYST_ID_BTN_ADDR');
+        PSConfiguration::deleteByName('FC_OYST_ID_SMART_BTN_ADDR');
+        PSConfiguration::deleteByName('FC_OYST_POSITION_BTN_ADDR');
         // Conf 1-Click carriers
         PSConfiguration::deleteByName('FC_OYST_SHIPMENT_DEFAULT');
         PSConfiguration::deleteByName('OYST_ONE_CLICK_CARRIER');
@@ -336,7 +349,9 @@ class InstallManager
         PSConfiguration::deleteByName('FC_OYST_OC_REDIRECT_CONF');
         PSConfiguration::deleteByName('FC_OYST_OC_REDIRECT_CONF_CUSTOM');
         PSConfiguration::deleteByName('FC_OYST_MANAGE_QUANTITY');
-        PSConfiguration::deleteByName('FC_OYST_SHOULD_AS_STOCK');
+        if (PSConfiguration::hasKey('FC_OYST_SHOULD_AS_STOCK')) {
+            PSConfiguration::deleteByName('FC_OYST_SHOULD_AS_STOCK');
+        }
         PSConfiguration::deleteByName('FC_OYST_MANAGE_QUANTITY_CART');
         PSConfiguration::deleteByName('FC_OYST_ONLY_FOR_IP');
         // Conf 1-Click restrictions
